Initial Tasks
The first task that was undertaken was to create the user interface of the plugin. This involved creating selection boxes for the input and output parameters as well as a slider to control the stereo width.

Initial work was then put into developing the algorithms for the encoding and decoding of stereo and mid-side inputs for the plugin.

Once these algorithms were calcualted they were combined in an if statement to be used with the correct input and output as selected by the combination boxes.