Code Update and Phase Inversion
After the previous commit, it was found that when opening the plugin within Reaper no clear audio effects were taking place. After referring back to the code it was found that the problem existed in that the output processing of each option was infront of the stereo width effects in the signal chain. After switching the width and output equations around the stereo plugin worked fine.

Further work was put into adding more aspects to the plugin. A phase inversion option is in the works, all the relevant code is in the processor however there are no clear audio effects within Reaper. Further troubleshooting is taking place in order to find out what the problem is.

Research has also been had into adding other apsects to the plugin including; spectrum analyzer, LFO source for modulation, panning effects as well as introducing a more novel aspect to the plugin.